+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/structure/building/function/Research.java
@@ -122,7 +122,7 @@ implements Lab, Serializable {
 
             int techLevel = buildingConfig.getResearchTechLevel(buildingName);
             int labSize = buildingConfig.getResearchCapacity(buildingName);
-            double buildingResearchSupply = techLevel * labSize;
+            int buildingResearchSupply = techLevel * labSize;
 
             result += buildingResearchSupply * existingResearchValue;
         }
