+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/structure/building/function/WasteDisposal.java
@@ -91,7 +91,7 @@ public class WasteDisposal extends Function implements Serializable {
 
 			int techLevel = buildingConfig.getWasteDisposalTechLevel(buildingName);
 			int labSize = buildingConfig.getWasteDisposalCapacity(buildingName);
-			double buildingWasteDisposalSupply = techLevel * labSize;
+			int buildingWasteDisposalSupply = techLevel * labSize;
 
 			result += buildingWasteDisposalSupply * existingWasteDisposalValue;
 		}
